Add logback & Slack Api
- Write log to file by lockback
- Push exception/ error to Slack by Slack Api

+++ CMS/src/main/java/edu/funix/common/ICategoryService.java
@@ -11,7 +11,7 @@ public interface ICategoryService {
     
     List<CategoryModel> search(PageModel page, String searchKey) throws SQLException, Exception;
 
-    CategoryModel findCategoryById(int id) throws SQLException, Exception;
+    CategoryModel findCategoryById(long id) throws SQLException, Exception;
 
     void save(CategoryModel category) throws SQLException, Exception;
